requirements.txt
python-dotenv
watchdog
customtkinter
speechrecognition
pyaudio  # Eğer PyAudio kurulumunda sorun yaşarsanız, alternatif olarak sounddevice deneyebilirsiniz.
Pillow
langchain-google-genai
langchain
pyttsx3 